Use default globing for .ts files
Less configuration and overhead.

From the docs:
A) “If no "files" property is present in a tsconfig.json, the compiler
defaults to including all TypeScript (*.ts or *.tsx) files in the
containing directory and subdirectories.”
B) “The "files" property cannot be used in conjunction with the
"exclude" property. If both are specified then the "files" property
takes precedence.”
http://www.typescriptlang.org/docs/handbook/tsconfig-json.html